% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_ppts_interp.R
\name{ff_ppts_interp_linear}
\alias{ff_ppts_interp_linear}
\title{Linearly interpolate missing data in-between dates with observed data}
\usage{
ff_ppts_interp_linear(
  df_data,
  ar_svr_group = c("location", "variable"),
  svr_data = c("population"),
  svr_date = c("year"),
  svr_interp = c("population_interp1"),
  verbose = FALSE
)
}
\arguments{
\item{df_data}{data frame with various time-series/panel with missing data}

\item{ar_svr_group}{array of grouping variable names, panel groups}

\item{svr_data}{string variable name for data variable}

\item{svr_date}{string variable name for date variable}

\item{svr_interp}{string variable name for variable with interpolated values}

\item{verbose}{Boolean print details}
}
\value{
A dataset with interpolated values
}
\description{
A file, \code{df_data}, has grouping variables \code{ar_svr_group}
and a data variable \code{svr_data} (e.g., population), measured at sequential
\code{svr_date} (e.g., year, month) dates. There are nonconsecutive gaps in \code{svr_data}
measurements, with missing values in some dates. We use the rate of change
between the closest dates in which there is data to linearly interpolate
\code{svr_data} in dates with missing values. The output is stored in variable
\code{svr_interp}.
}
\references{
\href{https://github.com/FanWangEcon/PrjCompPPTS/issues/16}{PrjThaiHFID-issue-16}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
