% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_ppts_lrce.R
\name{ff_ppts_lrce_fel}
\alias{ff_ppts_lrce_fel}
\title{Generate FEL file}
\usage{
ff_ppts_lrce_fel(df_fpc, verbose = FALSE)
}
\arguments{
\item{df_fpc}{Dataframe input, output of \code{PrjCompPPTS::ff_ppts_lrce_fpc}
\href{https://github.com/FanWangEcon/PrjThaiHFID/issues/1}{PrjThaiHFID-issue-1}.}

\item{verbose}{Boolean print details}
}
\value{
A dataset with ratios, levels, percentage changes, and elasticities.
}
\description{
Generate elasticities
}
\references{
\href{https://github.com/FanWangEcon/PrjCompPPTS/issues/7}{PrjThaiHFID-issue-7, Step 3}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
