% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_ppts_lrce.R
\name{ff_ppts_lrce_fpc}
\alias{ff_ppts_lrce_fpc}
\title{Generate FPC file}
\usage{
ff_ppts_lrce_fpc(
  df_flr,
  ar_it_years_chg = c(1960, 1980, 2000, 2020),
  ls_chg_years = list(chg_80v60 = c(1960, 1980), chg_00v80 = c(1980, 2000), chg_20v00 =
    c(2000, 2020), chg_20v80 = c(1980, 2020)),
  verbose = FALSE
)
}
\arguments{
\item{df_flr}{Dataframe input, output of \code{PrjCompPPTS::ff_ppts_lrce_flr}}

\item{ar_it_years_chg}{Array of integer, years to select for computing changes}

\item{verbose}{Boolean print details}

\item{ls_chg_year}{List of Array each array with start and end date and list contains
new variable names}
}
\value{
A dataset with ratios, levels, and percentage changes.
}
\description{
Compute percentages changes across time
}
\references{
\href{https://github.com/FanWangEcon/PrjCompPPTS/issues/7}{PrjThaiHFID-issue-7, Step 2}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
