% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_ppts_interp.R
\name{ff_ppts_pchg}
\alias{ff_ppts_pchg}
\title{Compute per-period (e.g., annualized) percentage changes}
\usage{
ff_ppts_pchg(
  df_data,
  ar_svr_group = c("location", "variable"),
  svr_data = c("population"),
  svr_date = c("year"),
  svr_pchg = c("population_pchg_yr"),
  verbose = FALSE
)
}
\arguments{
\item{df_data}{data frame with various time-series/panel with missing data}

\item{ar_svr_group}{array of grouping variable names, panel groups}

\item{svr_data}{string variable name for data variable}

\item{svr_date}{string variable name for date variable}

\item{svr_pchg}{string variable name for variable with storing the per-period
percentage change (e.g., annualized changes)}

\item{verbose}{Boolean print details}
}
\value{
A dataset with interpolated values
}
\description{
A file, \code{df_data}, has grouping variables \code{ar_svr_group}
and a data variable \code{svr_data} (e.g., population), measured at sequential
\code{svr_date} (e.g., year, month) dates. There might be non-consecutive gaps in
\code{svr_data} measurements, with missing values in some dates. We compute per-period
changes, annualized if the date variable is yearly, between all dates when data
is available.
}
\references{
\href{https://github.com/FanWangEcon/PrjCompPPTS/issues/17}{PrjThaiHFID-issue-17}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
