% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppts_country_code}
\alias{ppts_country_code}
\title{Country, region, and subcountry code and names}
\format{
A data frame with 266 rows and 7 variables:
\describe{
\item{location_name}{Location full name}
\item{location_code}{Short string code for country and regions}
\item{location_region_group}{World bank seven regions}
\item{location_income_group}{World bank country income groups}
\item{location_wblend_group}{World bank lending category}
\item{location_region_group_code}{3 Let. WB 7 region + digit sorter}
\item{location_code_adj}{Allow for region then country by region sort}
}
}
\source{
World bank
}
\usage{
ppts_country_code
}
\description{
Country, region, and subcountry code and names
}
\examples{
data(ppts_country_code)
summary(ppts_country_code)
print(ppts_country_code)
}
\keyword{datasets}
