% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppts_easia_weuro_world_elas_interp1}
\alias{ppts_easia_weuro_world_elas_interp1}
\title{Elasticities across all potential variable pairs and for different time frames, non-interpolated}
\format{
A data frame:
\describe{
\item{location_code}{Short string code for country and regions}
\item{location_level}{Country, or regional or sub-country}
\item{variable}{Statistical variable}
\item{year_bins_type}{The type of year binning}
\item{year_bins}{Year bins given year binning}
\item{pchg_interp1}{Percentage change from start to end of bin associated with variable (for the denominator), with interpolation and extrapolation}
\item{value_interp1}{Level values for variables, with interpolation and extrapolation}
\item{variable_numerator}{The variable, whose percentage change is computed, in the numerator}
\item{pchg_interp1_numerator}{\code{pchg} but for the \code{variable_numerator} variable, with interpolation and extrapolation}
\item{elasticity_interp1}{\code{pchg_interp1_numerator} over \code{pchg_interp1} to compute elasticities over different windows for all possible variable \code{variable_numerator} and \code{variable} combinations.}
}
}
\source{
World bank, national statistics websites of W. Europe and E. Asia countries
}
\usage{
ppts_easia_weuro_world_elas_interp1
}
\description{
Compute elasticities for different numerator and denominator combinations, using non-interpolated data.
}
\examples{
data(ppts_easia_weuro_world_elas_interp1)
summary(ppts_easia_weuro_world_elas_interp1)
print(ppts_easia_weuro_world_elas_interp1)
}
\keyword{datasets}
