% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppts_easia_weuro_world_elas_raw}
\alias{ppts_easia_weuro_world_elas_raw}
\title{Elasticities across all potential variable pairs and for different time frames, non-interpolated}
\format{
A data frame:
\describe{
\item{location_code}{Short string code for country and regions}
\item{location_level}{Country, or regional or sub-country}
\item{variable}{The variable associated with \code{value}, also the variable, whose percentage change is computed, in the denominator}
\item{year_bins_type}{The type of year binning, 1 year, 10 years, 20 years, etc.}
\item{year_bins}{Year bins given year binning, labels of different years, decades, bi-decades, etc.}
\item{pchg}{Percentage change from start to end of bin associated with variable (for the denominator)}
\item{value}{Level values for variables}
\item{variable_numerator}{The variable, whose percentage change is computed, in the numerator}
\item{pchg_numerator}{\code{pchg} but for the \code{variable_numerator} variable}
\item{elasticity}{\code{pchg_numerator} over \code{pchg} to compute elasticities over different windows for all possible variable \code{variable_numerator} and \code{variable} combinations.}
}
}
\source{
World bank, national statistics websites of W. Europe and E. Asia countries
}
\usage{
ppts_easia_weuro_world_elas_raw
}
\description{
Compute elasticities for different numerator and denominator combinations, using non-interpolated data.
}
\examples{
data(ppts_easia_weuro_world_elas_raw)
summary(ppts_easia_weuro_world_elas_raw)
print(ppts_easia_weuro_world_elas_raw)
}
\keyword{datasets}
