% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppts_easia_weuro_world_pchg}
\alias{ppts_easia_weuro_world_pchg}
\title{Level and changes, interpolated and extraplated}
\format{
A data frame with 266 rows and 7 variables:
\describe{
\item{location_code}{Short string code for country and regions}
\item{location_level}{Country, or regional or sub-country}
\item{variable}{Statistical variable}
\item{year_bins_type}{The type of year binning}
\item{year_bins}{Year bins given year binning}
\item{value}{Level at start to end of bin}
\item{value_interp1}{\code{value} with interpolation and extrapolation}
\item{pchg}{Percentage change from start to end of bin}
\item{pchg_interp1}{\code{pchg} with interpolation and extrapolation}
}
}
\source{
World bank, national statistics websites of W. Europe and E. Asia countries
}
\usage{
ppts_easia_weuro_world_pchg
}
\description{
Level and changes, interpolated and extraplated
}
\examples{
data(ppts_easia_weuro_world_pchg)
print(ppts_easia_weuro_world_pchg)
}
\keyword{datasets}
