% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffp_ppts_lrce.R
\name{ff_ppts_lrce_flr}
\alias{ff_ppts_lrce_flr}
\title{Generate FLR file}
\usage{
ff_ppts_lrce_flr(
  df_ppts,
  st_year_bins_type = "1940t2020i01",
  ar_it_years = c(1960, 1980, 1990, 2000, 2010, 2020),
  verbose = FALSE
)
}
\arguments{
\item{df_ppts}{Dataframe input.}

\item{st_year_bins_type}{String category string}

\item{ar_it_years}{Array of Integers subset of year bins to select}

\item{verbose}{Boolean print details}
}
\value{
A dataset with ratios and levels.
}
\description{
Generate ratios
}
\references{
\href{https://github.com/FanWangEcon/PrjCompPPTS/issues/7}{PrjThaiHFID-issue-7, Step 1}
}
\author{
Fan Wang, \url{http://fanwangecon.github.io}
}
